EUPS packaging and bin/firefly
Add packaging files for EUPS build and distribution system, and a script --
bin/firefly -- designed to be invoked by the end-user to start firefly as a
desktop-style application.

Once the 'firefly' executable is run (it's in bin/, and EUPS' setup will add
$FIREFLY_DIR/bin to the path), it will start Tomcat with the fftools webapp
on a randomly chosen port.  It will print out the port on which it's
running, and the browser may be pointed there (or, alternatively, the Python
wrappers could in the future parse the result to communicate with the backend).

If one wants to define the ports instead of having them chosen at random
(useful when debugging, or when not working on a shared machine), define the
FFPORTS environmental variable, e.g.:

    FFPORTS="8080 8005" firefly

The http server will run on the first port, and the second port is the
"control" (aka shutdown) port.

Implementation: Firefly installed via EUPS builds the webapp and installs
the .war file, the unpacked webapp, and the 'bin/firefly' script into its
install directory.  Every time the bin/firefly script is run it will create
a custom (throwaway) CATALINA_BASE directory (in $TMPDIR) into which it
links the unpacked webapp.  It then starts Tomcat with the randomly chosen
ports.  That allows the Firefly and Tomcat EUPS-controlled install
directories to be read-only, while enabling having multiple instances of
Tomcat running on the same machine, and running the same app, w/o
collisions.

Running "in-tree": firefly script will detect when it's run from the source
directory (i.e., in the directory that was cloned with git).  Instead of
creating a throwaway CATALINA_BASE, it will use the one in './tomcat'
subdirectory.  That directory has the .war file linked to
../../build/libs, so it will automatically pick it up any time its rebuild.
This should enable this type of edit-build-test loop:

   gradle :fftools:war   # build the library

   firefly &   # Start firefly
   ... connect to it in the web browser, or otherwise ...

   ... make some changes to the source code ...
   gradle :fftools:war  # rebuild

Note: Tomcat should automatically detect that the war changed and restart
the webapp.